setup: exclude celery 4.4.4 which is broken due to unexpressed dependency
Celery 4.4.4 introduced the use of the 'future' package but forgot to
express it in its dependencies.

We could add the missing dependency on 'future' in Kallithea, but since the
problem is already fixed upstream shortly after 4.4.4 was released [1], we can
be sure that the next release (presumably 4.4.5) will contain the fix.

[1] https://github.com/celery/celery/pull/6146