middleware: drop sentry support
Our sentry support didn't seem to work with py3. It doesn't seem to ever have
been tested on py3 - it failed already when installing.

It seems like we used the 'legacy client':
  https://docs.sentry.io/clients/python/
and nowadays one is expected to use:
  https://docs.sentry.io/platforms/python/
which has integrations for Pyramid and other frameworks (but not TG2). There
are Celery and SQLAlchemy integrations too. See also the migration guide:
https://docs.sentry.io/platforms/python/migration/ .

Also, there is not much essential code in the old sentry support, and it seems
like it would be easier to reimplement from scratch. There is thus not much
lost by dropping it.