docs: drop documentation for installing without virtualenv
Installing globally as root require an unjustified trust in our build/install
process ... and makes it very hard to uninstall or clean up.

Installing as a user is slightly better but also more error-prone and not
recommended.

Those who really want to do it and know how to do it also don't need this
description.