autocomplete: also match on email address
There are cases where the firstname/lastname known to Kallithea, is not
complete, perhaps only an initial. But users that try to add a reviewer, may
start typing a full name and expect autocompletion.

For example:

    firstname: John
    lastname: D
    email: john.doe@example.com

When a review starts typing 'Doe'  they will not get any matches if
autocompletion is purely based on names.

This situation sometimes arises with Indian developers, where the names
known to companies may be abbreviated as initials. For example "C K Deepak"
where Deepak is the first name and C K are the initials of the last name.