autocomplete: also query 'firstname lastname' and 'lastname firstname' combinations
The autocomplete functionality for user names, e.g. in pull request reviewer
lists, @mentions, etc. would match the input term only on firstname,
lastname or username, but not a combination of firstname lastname.

This is a problem when there are many matches on the same firstname or
lastname, in particular with Chinese names like 'Wang', 'Cheng', etc. If you
know the full name and type it, you would not get any matches.

Instead, adapt the queries to also match on 'firstname lastname' and
'lastname firstname'.

This means that simple matching on only username or only lastname, can be
removed.