api: stop using 'Optional', 'OAttr'/'OptionalAttr' classes
There does not seem to be a good reason to use the 'Optional' and
'OptionalAttr' classes.
It makes the code harder to understand. And worse, the 'default value'
specified is not always used, which can thus give false information to
users.

The way Optional was used in the API calls is twofold:

1.either by effectively extracting a value, via Optional.extract(param).
  If 'param' was indeed specified by the user, then this would yield that
  user-specified value. Otherwise, it would yield the value declared in the
  parameter declaration, e.g. param=Optional(defaultvalue).

2.or by checking if a variable is an instance of the Optional class. In case
  a user effectively passed a value, this value will not be of the
  Optional class. So if a parameter is an object of class Optional, we know
  the user did not pass a value, and we can apply some default.

In the declaration of the parameter, the specified default value will only
be used if the 'extract' method is used, i.e. method 1 above.

A simpler way to address this problem of default values is just with Python
default values, using 'None' as magic value if the default will be
calculated inside the method.

The docstrings still specify something like:
   type: Optional(bool)
which is humanly readable and does not necessarily refer to a class called
'Optional', so such strings are kept.