model: stop using 'Optional' class in Setting.create_or_update()
The 'Optional' class can help to identify whether a value is set by the
caller or a default value.

But if you can make the assumption that the caller will never specify a
'None' value as 'real' value, then the Optional class is not really
necessary.

As this usage in the Setting class is the only remaining usage of
'Optional', it does not seem worth it to keep it.